backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-multipart==0.0.12
websockets==13.1
python-dotenv==1.0.1
